Kahe paketi kõrvutamine päeva lõikes(kuu ja aasta tulevad varsti)

+++ GraafikControl/Graafik.vb
@@ -2,11 +2,11 @@
     Public Sub ClearPoints()
         HinnaGraafik.Series("Pakett 1").Points.Clear()
     End Sub
-    Public Sub setPoint1(ByVal Time As String, ByVal hind As Double)
+    Public Sub setPoint1(ByVal Time As String, ByVal hind As Decimal)
         HinnaGraafik.Series("Pakett 1").Points.AddXY(Time, hind)
     End Sub
 
-    Public Sub setPoint2(ByVal Time As String, ByVal hind As Integer)
+    Public Sub setPoint2(ByVal Time As String, ByVal hind As Decimal)
         HinnaGraafik.Series("Pakett 2").Points.AddXY(Time, hind)
     End Sub
